Added primitives and started runtime frames

+++ src/type.h
@@ -5,17 +5,19 @@
 #ifndef ILLUMINA_TYPE_H
 #define ILLUMINA_TYPE_H
 
+#include <stdint.h>
+
 typedef struct TypeObject {
-    unsigned int size;
+    uint16_t size;
 } TypeObject;
 
 #define OBJECT_PROPERTY \
-        unsigned int reachable; \
+        uint8_t reachable; \
         TypeObject *type;
 
 #define INIT_OBJ(type) 0, type;
 
 #define OBJECT_PROPERTY_SIZE \
-        sizeof(unsigned int) + sizeof(TypeObject)
+        sizeof(uint8_t) + sizeof(TypeObject)
 
 #endif //ILLUMINA_TYPE_H
